refactor: cleanup AvatarList, MetaData, Tabs and Toolbar components
