+++ src/components/MetaData/MetaDataItem/MetaDataItem.scss
@@ -1,0 +1,22 @@
+@import '../../../styles/settings/colors';
+
+/* ==========================================================================
+   Component: Meta data item
+   ========================================================================== */
+
+.c-meta-data__item {
+	margin-right: 0.8rem;
+	font-size: 1.3rem;
+	color: $color-gray-400;
+}
+
+.c-meta-data-item--icon {
+	display: flex;
+	align-items: center;
+
+	.o-svg-icon {
+		width: 1.4rem;
+		height: 1.4rem;
+		margin-right: .4rem;
+	}
+}
