+++ src/components/Tabs/Tabs.scss
@@ -1,16 +1,10 @@
+@import '../../styles/settings/colors';
+
 /* ==========================================================================
    Component: Tabs
    ---
    Top navigation (usually inside of a toolbar)
    ========================================================================== */
-
-/* Variables
-   ========================================================================== */
-
-$c-tabs-item-color: $color-gray-300 !default;
-$c-tabs-item-active-color: $color-teal-bright !default;
-$c-tabs-item-border-color: $color-gray-200 !default;
-$c-tabs-item-hover-border-color: $color-gray-200 !default;
 
 /* Component
    ========================================================================== */
@@ -26,40 +20,6 @@ $c-tabs-item-hover-border-color: $color-gray-200 !default;
 	justify-content: center;
 }
 
-.c-tabs__item {
-	color: $c-tabs-item-color;
-	padding: 1.2rem;
-	font-weight: 500;
-	transition: color 0.1s linear;
-
-	display: flex;
-	align-items: center;
-	text-decoration: none;
-
-	.o-svg-icon {
-		margin-right: 1rem;
-		width: 1.8rem;
-		height: 1.8rem;
-	}
-
-	&:hover {
-		color: $c-tabs-item-active-color;
-	}
-
-	&:focus {
-		color: $c-tabs-item-active-color;
-	}
-
-	&.c-tabs__item--active {
-		box-shadow: inset 0 -0.2rem 0 $color-primary;
-		margin-bottom: -0.1rem;
-		color: $c-tabs-item-active-color;
-		font-weight: $g-headings-font-weight;
-		position: relative; // makes box-shadow to be on top of the global border
-		z-index: $g-c-tabs-z;
-	}
-}
-
 .c-tabs.c-tabs--bottom-border:after {
 	content: "";
 	display: block;
@@ -68,9 +28,4 @@ $c-tabs-item-hover-border-color: $color-gray-200 !default;
 	box-shadow: inset 0 -0.1rem 0 $color-gray-200;
 	position: absolute;
 	bottom: 0;
-}
-
-.c-tabs__item-length {
-	font-weight: 400;
-	margin-left: 0.4rem;
 }
