+++ src/components/Tabs/Tabs.tsx
@@ -5,6 +5,8 @@ import classNames from 'classnames';
 import { DefaultProps } from '../../types';
 import { Tab, TabProps } from './Tab/Tab';
 
+import './Tabs.scss';
+
 export interface TabsProps extends DefaultProps {
 	tabs: TabProps[];
 	border?: boolean;
