+++ src/components/Toolbar/Toolbar.test.tsx
@@ -3,7 +3,7 @@ import React, { Fragment } from 'react';
 
 import { Toolbar } from './Toolbar';
 import { ToolbarCenter, ToolbarLeft, ToolbarRight } from './Toolbar.slots';
-import { ToolbarItem } from './ToolbarItem';
+import { ToolbarItem } from './ToolbarItem/ToolbarItem';
 
 describe('<Toolbar />', () => {
 	it('Should be able to render', () => {
