+++ src/index.ts
@@ -12,7 +12,7 @@ import { useKeyPress } from './hooks/useKeyPress';
 import { Alert } from './components/Alert/Alert';
 import { Avatar } from './components/Avatar/Avatar';
 import { AvatarIcon } from './components/Avatar/AvatarIcon/AvatarIcon';
-import { AvatarList } from './components/Avatar/AvatarList/AvatarList';
+import { AvatarList } from './components/AvatarList/AvatarList';
 import { Badge } from './components/Badge/Badge';
 import { Blankslate } from './components/Blankslate/Blankslate';
 import { Box } from './components/Box/Box';
@@ -101,8 +101,8 @@ import { Toggle } from './components/Toggle/Toggle';
 import { ToggleButton } from './components/ToggleButton/ToggleButton';
 import { Toolbar } from './components/Toolbar/Toolbar';
 import { ToolbarCenter, ToolbarLeft, ToolbarRight } from './components/Toolbar/Toolbar.slots';
-import { ToolbarItem } from './components/Toolbar/ToolbarItem';
-import { ToolbarTitle } from './components/Toolbar/ToolbarTitle';
+import { ToolbarItem } from './components/Toolbar/ToolbarItem/ToolbarItem';
+import { ToolbarTitle } from './components/Toolbar/ToolbarTitle/ToolbarTitle';
 import { Tooltip } from './components/Tooltip/Tooltip';
 import { TooltipContent, TooltipTrigger } from './components/Tooltip/Tooltip.slots';
 import { WYSIWYG } from './components/WYSIWYG/WYSIWYG';
